Chapter 3 の142ページまで終わりま。

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -5,14 +5,19 @@ namespace App\Http\Controllers\Tweet;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Tweet;
+use App\Service\TweetService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
     /**
      * Handle the incoming request.
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService  $tweetService)
     {
+        if(!$tweetService->checkOwnTweet($request->user()->id, $tweetId)){
+            throw new AccessDeniedHttpException('You do not have permission to delete this tweet');
+        }
         $tweetId = (int) $request->route('tweetId');
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
